rename: 공통 컴포넌트 폴더 생성 후 옮김
