+++ src/components/rates/popcornRate/PopcornRate.style.ts
