+++ src/components/rates/reviewRate/ReviewRate.style.ts
