+++ src/components/rates/screeningRate/ScreeningRate.style.ts
